requirements.txt
llama-cpp-python
fastapi
uvicorn
langchain
chromadb
pymupdf
sentence-transformers